+++ src/components/filterBar/FilterBar.tsx
@@ -1,0 +1,38 @@
+"use client";
+
+import React from "react";
+
+interface FilterBarProps {
+  categories: string[];
+  selectedCategory: string;
+  onCategoryChange: (category: string) => void;
+}
+
+const FilterBar: React.FC<FilterBarProps> = ({
+  categories,
+  selectedCategory,
+  onCategoryChange,
+}) => {
+  return (
+    <div className="flex justify-center mb-8 overflow-x-auto py-2">
+      <div className="flex space-x-2">
+        {categories.map((category) => (
+          <button
+            key={category}
+            onClick={() => onCategoryChange(category)}
+            className={`px-4 py-2 rounded-full text-sm font-medium transition-colors
+              ${
+                selectedCategory === category
+                  ? "bg-blue-600 text-white"
+                  : "bg-gray-100 text-gray-800 hover:bg-gray-200"
+              }`}
+          >
+            {category.charAt(0).toUpperCase() + category.slice(1)}
+          </button>
+        ))}
+      </div>
+    </div>
+  );
+};
+
+export default FilterBar;
